Improve mobile experience and fix bugs
Disabled signup functionality for new users due to client request and
modified mobile table to be more user-friendly.

Auth logic was also reworked and now correctly works using both client-
side and ssr components. This is a complex implementation which is
correct and secure, but time inneficient, which creates consideration of
only using client-side or ssr authentication in the future, unless a mix
of both is desired.

Fixed false positive triggers of the confirmation/error toasters.
They now validate the data correctly and only trigger when a form is
submitted.

Next up: Implement complex animations using GSAP.